+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+* (Fixes https://github.com/vividus-framework/vividus-studio/issues/380) Add 'Insert Step' source code action
+
 ## [0.1.5] - 2023-02-23
 
 ### Fixed